+++ playlet-web/src/lib/Api/PlayletApi.ts
@@ -132,6 +132,10 @@ export class PlayletApi {
         return await fetch(`${PlayletApi.host()}/api/search-history`, { method: "DELETE" });
     }
 
+    static async clearCache() {
+        return await fetch(`${PlayletApi.host()}/api/cache`, { method: "DELETE" });
+    }
+
     static async getBookmarkFeeds() {
         const response = await fetch(`${PlayletApi.host()}/api/bookmarks/feeds`);
         return await response.json();